Optionally read fuzz target info from JAR manifest
By reading the fuzz target class and custom hooks from a JAR manifest
entry rather than a command-line argument, fuzz targets can be fully
self-contained.

This commit adds a Java function that looks for a unique
`Jazzer-Fuzz-Target-Class` attribute in all manifests on the classpath
and returns it to the driver if found. If no such entry is found, it
falls back to the `--target_class` commandline parameter.

In a similar way, the agent prepopulates the list of custom hooks to
load with the merged values of `Jazzer-Hook-Classes` attributes in all
manifests.