Skip instrumentation for known no-throw methods
A method invocation only needs to be instrumented for coverage if the
method can throw an exception (including RuntimeException, which is not
declared via `throws`). For methods in the Java standard library
(in `java.*`), this property is thoroughly documented via Javadoc.

This commit adds a doclet, i.e., a Java program that parses Javadoc, for
automatically compiling a list of known no-throw methods in `java.*`.
The list is loaded lazily at runtime and used to skip calls to these
methods in the coverage instrumentation.

With OpenJDK 15, the list consists of almost 9000 methods, many of which
are used very frequently and in performance-critical parts, e.g.,
StringBuilder#append(String) and List#size().