Increase coverage map size dynamically
As we are now instrumenting a known number of edges, we can start with
a small coverage map that grows as needed.

The initial size of the map is 512 and doubles whenever the instrumentor
runs out of new edge IDs.

As the bytecode instrumentation needs to have access to a contiguous
ByteBuffer, we allocate a (hopefully large enough for all practical
purposes) coverage counter buffer in native code once and only adjust
the size of the ByteBuffer dynamically. As libFuzzer is only made aware
of the part of the buffer in active use, this has no performance impact
and only very minor memory usage (< 10 MB).