Make value profile independent of Jazzer build
Currently, when using value profile, the behavior of the fuzzer on a
fixed fuzz target does not just depend on the seed, but also on the
particular binary build of the fuzzer (and thus on the compiler used).
This is because the trampoline that translates the deterministic fake
PCs injected by the bytecode into return addresses for the libFuzzer
hooks may be located at different addresses in the fuzzer binary.

This commit determines the address of the trampoline at startup and
adjusts all fake PCs accordingly before passing them into the
trampoline.