Make driver check for manually reported finding
Not all findings can easily be reported as uncaught exceptions, e.g.
exceptions raised in a hook might be caught and swallowed by library
code.

This change adds a global variable in the Java runtime that is checked
by the fuzzer after every invocation of the fuzz target. If it contains
a finding (represented as a Throwable), it is reported by Jazzer.