Add Jazzer API method for reporting findings from hooks
The method Jazzer.reportFindingFromHook can be used to reliably report a
finding from a hook, even if the hook is called within an overly generic
exception handler.

In order to make the fuzz target return quickly, a hard to catch error
is thrown after the finding has been registered via the global finding
variable. This is not guaranteed to terminate the execution of the fuzz
target right away, but the global finding variable ensures that the
finding will be reported eventually (unless the fuzz target never
returns).