Speculative fix for crashes in intercepted memcmp
Jazzer currently segfaults in __sanitizer_weak_is_relevant_pc when it is
called from an intercepted memcmp on a non-main JVM thread. This hints
at a thread-safety issue since the vector with address ranges to ignore
can be read while it is being modified.

With this commit, reads to the vector only occur after the only writes.