Fix: Insert method probes after invocations
Coverage probes for method invocations should be emitted after the
invocation as their purpose is to track whether the method has returned.
This is in line with how unpatched JaCoCo instrumentes lines with method
calls: It emits the probe when it visits the line number label, which
happens after all instructions associated with the line have been
processed.