Ensure default -Xmx value is below -rss_limit_mb
libFuzzer defaults to an rss_limit_mb of 2048, but we start the JVM with
-Xmx4096m. This can lead to libFuzzer OOM reports when a single
allocation in Java exceeds 2 GB but still fits into the JVM heap.

This is solved by letting the JVM heap size default to slightly less
than 2 GB.

This change is not fully backwards compatible, but will only cause
targets to crash more often than they used to.