Pass quoted arguments to child processes
libFuzzer does not quote the arguments it passes to child processes during merge
and fork, which leads to arguments being lost if passing multiple jvm_args with
delimiter ';'.

This commit adds a libFuzzer patch that properly quotes all arguments as well as
a test that fails if quoting is not appropriate.