Fix default method return value recording
Using a proxy to implement RecordingFuzzedDataProvider led to a subtle
bug related to default methods: Since the invocation handler cannot pass
the proxy instance itself to invoke (this would create a cycle), calls
to non-default methods that were internal to default methods were not
recorded as they occured on the original object.

There does not seem to be a way to fix this with proxies without either
making the FuzzedDataProvider interface aware of it being proxied (which
likely incurs a performance penalty while fuzzing) or duplicating
default method implementations.

A simple and less laborious fix is to implement the proxy object by hand
with every method delegating to the target and recording the return
value.