Do not shade soot.**
Soot refers to its own classes via string constants in a few places and
not all of them seem to be detected correctly by the shading tool. Since
not shading soot doesn't have many downsides - it is unlikely to be a
production dependency of another library we want to fuzz, don't shade it
for now.