Remove C++ dependencies on CoverageMap internals
This commit moves almost all coverage map logic into the Java class
with the aim of making it easier to replace the particular
implementation of the coverage map and reducing the amount of JNI code.

This is almost a pure refactoring, but also makes the allocations of
fake instructions and libFuzzer PC entries dynamic. This both slightly
reduces the constant memory footprint of the fuzzer and further
improves the encapsulation of the Java CoverageMap implementation as the
native code no longer has to know the maximum size of the map.