Fix TurbojpegFuzzer native library path
The native library path for this example got lost during the migration
to rules_jni and is required for the fuzz target to be functional.

Also updates rules_foreign_cc and cleans up the attributes passed to
the cmake rule. This includes adding -fPIC as a speculative fix for
relocation issues reported by the linker.