Use tags for all dependencies
GitHub support has confirmed that only the /archive/refs/tags endpoint
can be assumed to return stable results over time.

For dependencies that do not offer tags we can use, I created a fork
with a "commit-<SHA>" tag for the commit we were previously depending
on.

Along the way, update some dependencies to their latest tag.

See this comment for more context regarding the GitHub guarantees for
archive stability:
https://github.com/bazel-contrib/SIG-rules-authors/issues/11#issuecomment-1029861300