Enable BuildBuddy integration
Enable BuildBuddy integration when the API key is available in the
workflow. Forks should still be able to execute the workflow without
piping logs to BuildBuddy.

This is done by setting the environment variable BUILD_BUDDY_CONFIG
based on the API secret. The script echoBuildBuddyConfig returns an
environment config stating the needed bazel settings if the API token is
passed in as first argument, otherwise an empty string is returned. The
return value is used to update the workflow environment settings.

The passed around secret is filtered out in the GitHub logs.