Simplify registering PC tables
libFuzzer only checks the value of a PC when the corresponding flag is
not zero. Therefore, it is safe to set the all PC entries to any value
as long as the corresponding flag is set to zero. We set the value of
each PC to the index of the corresponding edge ID. This facilitates
finding the edge ID of each covered PC reported by libFuzzer.