Extract native fuzzer callbacks into a shared library
When looking up a JavaCritical_ function, the OpenJDK will first look up
the library that contained the corresponding Java_ function and then try
to dlopen that library. If the Java_ function is defined by the driver,
this dlopen call will fail as it can't dynamically load executables.
This seems to a bug as the usual Java_ lookup goes through more
complicated logic that properly deals with statically linked native
libraries.

We work around this issue by loading the fuzzer callbacks from an honest
shared library packaged into the agent JAR.

Since dynamic linking is difficult on Windows and we don't even know
whether CriticalJNINatives work there, we continue to statically link
the callbacks there.