Ensure that libFuzzer can start subprocesses in cifuzz/jazzer
libFuzzer uses system() to start subprocesses, which requires a shell
to exist at /bin/sh. The regular distroless images don't include a
shell, which breaks -fork and -minimize_crash.

Since we only need a very minimal shell, the :debug variants with their
busybox shell seem well suited for this purpose, with the only caveat
that we have to symlink the shell binary into the correct place.