Simplify the array/map length heuristic
This didn't really make sense before as it relied on checking how many
bytes the construction of the first element consumed, but that crucially
depended on how many bytes are left rather than saying something about
the intrinsic "size" of the constructed type.

Now, we use JOL to compute the shallow size of a class instance - the
idea being that this is a decent estimate for the number of bytes
consume will require to construct a non-trivial instance of this class.
The length of a structure of variable length (array, map) is now
consumed from the fuzzer input.

Also adds a missing factor of 2 to the remainingBuckets logic for maps.