all: Only append hooks to the bootstrap classpath
Previously, we appended the entire Jazzer agent with all its
dependencies to the bootstrap classpath, which resulted in a number of
classes being loaded both in the system and bootstrap class loader when
the agent wasn't initialized right after the JVM had been started.

With this commit, the agent JAR bundles a separate bootstrap JAR that
only includes the classes that strictly must be available on the
bootstrap class loader's classpath: Classes defining callbacks called by
our instrumentation as well as the native driver library which these
callbacks call into.

The new AgentInstaller class takes care of extracting and appending this
JAR to the bootstrap search path. For the sake of consistency, it does
this even if --nohooks is specified, but doesn't enable the agent in
that case.