driver: Fix DEDUP_TOKEN computation after Java rewrite
After the Java rewrite of FuzzTargetRunner, we no longer execute the
fuzz target from native code, which means that FuzzTargetRunner itself
will be part of the stack trace. Since we do not want DEDUP_TOKEN to
change if we make changes to Jazzer itself, we stop the computation
before the first frame that belongs to Jazzer but isn't a sanitizer.