driver: Ensure that JVM shutdown hooks are called when exiting
Previously, we used `_Exit` to exit on a finding in order to get around
libFuzzer's "fuzz target exited" message. This worked well in the native
driver, which still had a chance to properly shut down the JVM and
ensure that its shutdown hooks are called, but doesn't work with Java
FuzzTargetRunner running in the JVM it needs to shut down.

Instead, we now set libFuzzer's `::fuzzer::RunningUserCallback` to
`false` after the finding produced by the actual fuzz target has been
processed.

Since we can now rely on our own shutdown hooks being called reliably,
this also simplifies the logic for deleting the jazzer_bootstrap.jar and
calling fuzzerTearDown.