junit: Do not leak fuzzer startup into first seed test
The fuzzer startup output and setup time is no longer added to the
first seed tested in the regression test.

The fuzzer startup output is instead recorded on the
`@ParameterizedTest` container. Unfortunately, there doesn't seem to be
a way to include the setup time in that container - the container time
is simply the sum of the times recorded for the individual seeds.