bazel: Replace agent deploy JAR with stand-alone Jazzer JAR
The only difference in the resulting JAR is that we no longer include
the hooks part of the API in the non-bootstrap class path and thus have
to add an explicit runtime dependency in our fuzz target macro. This
change will not be user-facing: Our jazzer Maven module will depend on
the jazzer-api module.