junit: Move agent installation into RegressionTestArgumentProvider
As became clear from `NoClassDefFoundError`s caused by stricter jar
stripping introduced by follow-up commits, the
`TestInstancePreConstructCallback`, when applied to a
`ParameterizedTest`, actually runs *after* the argument discovery has
started.

In order to ensure the agent and thus the bootstrap jar are loaded as
early as possible, ensuring that e.g. `FuzzedDataProviderImpl` can find
the `UnsafeProvider`, we instead initialize the agent in
`provideArguments`.