deploy: Let jazzer depend on jazzer-api
Even though Jazzer only needs a proper subset of all classes offered by
jazzer-api, it would be bad style to have these classes included in two
jars on the classpath.

In the native launcher case, having two separate jars to add to the
class path would be annoying, so we instead include a "standalone"
deploy jar bundling both that will be part of the GitHub releases and
the native launcher's Bazel runfiles.