Fix sanitizers not being loaded with jazzer_standalone_deploy.jar
I forgot to add the deploy manifest line to
`jazzer_standalone_deploy.jar`. Since this is error-prone, use code
generation to load the list of built-in hooks instead in a single
location.

Also add a check to verify that built-in hooks are loaded. This would
have let existing tests catch this regression.