bazel: Use Xcode on macOS
Xcode is the most popular toolchain on macOS and, even with recent
versions of macOS, automatically picks up the macOS SDK sysroot without
the need for manual configuration via compiler and linker flags.

In addition, the Bazel-provided Xcode toolchain can cross-compile to
both arm64 and x86_64, which will be used to simplify the release
workflow in a subsequent change.