sanitizers: Add SSRF detector
Server Side Request Forgery (SSRF) attacks occur when requests to
destinations based on untrusted data are executed. This could lead to
exfiltration of sensitive data or exposure of internal services.

The sanitizer detects such attacks by checking if the fuzzer was able to
inject a honeypot host name into the requested destination. If this
is the case, it reports a high severity security issue.

Checks are implemented via `java.net.Socket.connect` and
`java.nio.channels.SocketChannel.connect` methods.