runtime: Use `Log` via reflection
The runtime is loaded by the bootstrap loader and thus would depend on a
duplicated `Log` class if it depended on it directly. Instead, use
reflection to access `Log` loaded by the system class loader.

Also moves the burden to log errors encountered by the `UnsafeProvider`
out of the utility class and into the caller.