all: Use logger class everywhere
This commit replaces all direct usages of `System.err` and `System.out`
with equivalent `Log` method calls. 

This change is meant to be mostly invisible to users. In some cases the
formatting is changed as we no longer manually wrap lines, which we
never did in a consistent manner. If we want to bring back line breaks,
we should implement this in `Log` instead in a follow-up PR.

In some cases unstructured logging to stdout is moved to stderr.

Fixes FUZZ-480