Fix Docker build and prevent unnecessary rebuilds
Due to a Bazel issue (https://github.com/bazelbuild/bazel/pull/17158),
the recent introduction of `universal_binary` ended up forking the
configuration for dependencies of the Jazzer launcher. As a result,
some Java targets were built twice and the Docker images failed to build
as the Jazzer standalone jar was no longer a dependency of the launcher
in the default configuration.

As a workaround and also because it prevents similar issues on a more
fundamental level, switch to the soon to be default improved config hash
algorithm.