agent: Ignore offline instrumentation
JaCoCo is able to offline instrument classes. It determines if classes,
which should be instrumented dynamically, are already instrumented by
checking the existence of the coverage data field, and throws an
exception if it's present.

Jazzer's `IProbeArrayStrategy` already does not add the coverage data
field. This PR also removes JaCoCo's check if the field is present, and
so enables fuzzing of offline instrumented classes.