mutation: Validate annotation usage on fuzz target parameters
Annotations such as `@NotNull` and `@InRange` only apply to particular
types and were silently ignored when applied on other types. This could
lead to subtle issues such as the user (or the author of this commit)
declaring a `@NotNull byte[]` parameter and being surprised to find that
the byte array is sometimes `null` - the annotation applies to the
primitive array component type rather than the array itself.

This commit adds an `@AppliesTo` meta-annotation that can be used to
describe the valid usage sites of an annotation. Valid uses are enforced
both in `ArgumentsMutator` and the `StressTest`.