mutation: Add support for map proto fields
While map fields in protos are represented as repeated message fields to
the reflection API, this "illusion" wasn't convincing enough for us to
support them automatically: There is a single class that represents all
MapEntry messages, even though there are many different message types
due to the possible key/value type combinations.

In addition, modelling a map as a list of entries means that some
mutations, e.g. duplicating a list entry, would silently not affect the
map.

Instead, we are now mapping a proto map field to an actual map to which
we apply our map mutator.