bazel: Use a hermetic Java toolchain compiling for Java 8 (#671)
This simplifies the setup for contributors as well as the release process. Since we can't use `--release 8`, we still need to check compatibility with Java 8 runtime libraries in our CI, which still uses the local JDK 8 to run tests against.

Also removes stamping from Maven builds as we already have the version number baked in.