autofuzz: Fix logs for bug detector findings
Autofuzz tries to print caught exceptions, but this fails in case of bug
detector findings, as the thrown exceptions are intentionally hard to
catch and rethrow themselves if possible.  Due to this behavior the
following line is potentially loged multiple times: `ERROR: Unexpected
exception encountered during autofuzz:`

In this case the exceptions are also only used to abort the current
invocation, logging of the actual findings is done by different
means. So it's save to ignore this kind of exceptions.