junit: Compute a default for `jazzer.instrument` based on the class path
The new heuristic walks the classpath directory and considers all
directories (but not jar files) to constitute the classes of the current
project. All subdirectories that include `.class` files are translated
into packages and all their subpackages are subject to instrumentation.

This is expected to give much better results than the previous
heuristic, which only considered the first two or three segments of the
test class package.