mutator: Don't cross over all the time
Custom cross over and custom mutate are the only mutators registered in
libFuzzer, hence, cross over is picked as often as mutate, which is way
too frequently. Without custom mutate, cross over would be picked out
among the default mutators, so ~1/12 of the time. This also seems too
much and is reduced at 1/100 in the Java part of the code and mutate is
used in the other cases.