all: Only load Opt in the system class loader
We can detect whether we are running on Android via system properties
and thus don't need to have this bit passed in via a flag.

As a result, the runtime no longer depends on Opt, which is thus no
longer loaded in the bootstrap class loader in addition to the system
class loader.