junit: Reduce JUnit dependency requirements
Since `jazzer-junit` does not shade its dependencies, it should depend
on the lowest possible versions that provide all required features to
maximize compatibility.

We keep a reference to the latest JUnit Jupiter version in the
springboot test to have coverage with both the oldest supported and the
most recent version.