bazel: Mark an intermediate `java_binary` targets as non-executable
Executable `java_binary` targets, i.e., those that can be `bazel run` or
used in build actions, require a Java runtime for the target platform.
Since there is no standalone Java runtime for Android, all Java targets
in the transitive closure of an `android_binary` must not be executable
for cross-compilation to work.

Note: Cross-compilation without flags is currently blocked on
https://github.com/bazelbuild/bazel/issues/17085.