bazel: Update rules_jni
In version 0.8.0, `libjvm` has an optional instead of a mandatory
dependency on a Java runtime. This is necessary for builds on macOS
where we build `//launcher:jazzer` as a `universal_binary`, which
performs a split transition to both macOS amd64 and macOS arm64.
Whatever the host architecture is, the local JDK used during CI builds
will not be compatible with both architectures.