BREAKING: Accept more truthy values for options and reject invalid ones
All of `true`, `on`, `yes`, `y`, and `1` are now recognized as boolean
values indicating `true` when passed to boolean options. If the value of
a boolean option is neither of those nor any of the corresponding
"falsy" values, an error is reported instead of silently accepting the
value and interpreting it as `false`.

Note: This change is mildly breaking as it should only break fuzzer
invocations that weren't behaving as intended.