tests: Add benchmarks
The benchmark runs a fuzz target a given number of times with different
seeds and computes statistics (min, max, median, average) over the
number of runs required to produce the desired finding.

Benchmarks double as CI tests that verify that the given maximum number
of runs isn't exceeded. These tests will be enabled in a follow-up
commit.