mutation: Fix cross-over insertion logic
The implementations of the "insert chunk" cross-over actions for `List`
and `Map` incorrectly assumed that the chunk to be inserted would fit
into the collection to be modified, but this is only true for the other
actions.

This is fixed by inlining the helper methods used for the insert action
and making the required changes. As a side effect, we also get rid of a
redundant iteration over the "other" collection.