mutation: Use `hasFixedSize` in `sizeInClosedRange`
The benchmarks in `//tests/benchmarks` show that biasing the size of
subsets of collections of primitives chosen by the mutator to be small
results in much worse performance than a comparable unstructured fuzz
test.

Before this change, 11 out of 15 runs time out with no run limit, the
other ones result in:

```
{
  "values": [
    11143,
    28128,
    581194,
    4229980
  ],
  "minimum": 11143,
  "maximum": 4229980,
  "average": 1212611.25,
  "median": 304661
}
```

After this change, all runs pass within a limit of 35,000 runs:

```
{
  "values": [
    887,
    1557,
    1889,
    2557,
    3023,
    3346,
    3517,
    6075,
    6613,
    7991,
    9578,
    10850,
    15583,
    23638,
    31046
  ],
  "minimum": 887,
  "maximum": 31046,
  "average": 8543.333333333334,
  "median": 6075
}
```

ExperimentalMutatorComplexProtoFuzzer now takes more runs on Linux, but
still less than on other platforms, which seems to indicate that the
Linux seed just happened to be a lucky choice.