tests: Do not retain values in StressTest
StressTest required a lot of memory since it collected all `init` and
`mutate` return values in a list. Instead, cross values off of a short
list for "contains" type checks and use `hashCode()` to stand in for the
actual value in statistical tests.

Verified locally that the test now passes with `--jvmopt=-Xmx512M`.