mutation: Support lazy computation of `hasFixedSize`
This allows mutators for potentially recursive structures to still
derive their value of `hasFixedSize` from that of child mutators, which
is necessary to support mutators with arbitrary child mutators.

Mutators can either override `hasFixedSize` to return a fixed value or
`computeHasFixedSize` to compute and cache the value the first time
`hasFixedSize` is invoked.

This fixes a bug that previously resulted in `Any` messages always being
treated as variable size even with an `@AnySource` pointing to
exclusively fixed-size messages.