bazel: Always use the hermetic toolchain on Linux
The main reason for not using the hermetic toolchain has been that the
large one-time downloads of LLVM and sysroot could scare off
contributors. Now that we have gone private, this is no longer a concern
and the added reproducibility wins.

Also enable `layering_check` in all builds: It is currently unsupported
by the hermetic toolchain, but should be supported by all relevant
versions of Xcode.