feat(mutator): activate mutator framework by default
Use the mutator framework by default, if not only a single byte[] or
FuzzedDataProvider parameter is requested. If that's the case, still pass in
libfuzzer's input data directly, so that existing corpus entries don't break.

Code with the old approach still remains, as it should be possible to
revert back to the previous functionality. To do so set the "mutator_framework"
option to "false".

BREAKING CHANGE: Experimental mutator parameters are renamed and no longer
usable.
- "experimental_mutator" (default "false") -> "mutator_framework" (default
"true)
- "experimental_cross_over_frequency" (default 100) ->
"mutator_cross_over_frequency" (default 100)