fix: make sure ConsumeIntegralInRange is always in range [min; max]
For cases where FuzzedDataProvider.consume<Byte,Short,Char,Integer,Long>(from,
to) sets the range [from; to] to be exactly the maximum value for the type, this
methods could return values outside the requested range. This PR makes sure that
the returned values are always in range.

In addition, this PR boost the performance of these methods by making less use
of the modulo operator to reduce fuzzer values into the range.  Keeping the
number "as is" allows the fuzzer to use the table of recent compares more
efficiently. At the same time, this change might invalidate existing crash files
and some corpus entries.  However, given overall performance boost this change
brings, the potential corpus invalidation is considered less important.