fix(mutation): respect maxLength in LibFuzzerMutatorFactory.readExclusive
Since users are free to change fuzz test signature, readExclusive might return
byte arrays longer than maxLength. Calling libfuzzer's Mutate with maxSize that is less
than the length of the array to be mutated might result in a segfault at the time when the
data type is used in Java. Some of libfuzzer's mutators (Mutate_AddWordFromTORC) are not
designed to handle this case and will mutate past the array bounds.

Co-authored-by: Simon Resch <simon.resch@code-intelligence.com>