fix: don't apply Bean factories for proto enums
Using the full original factory to construct a mutator for protobuf enum
types would trigger a deeply nested mutator construction attempt on the
EnumValueDescriptor class using e.g. the
`CachedConstructorMutatorFactory`. While this still ended up at the
dedicated enum mutator for most cases, for complex protobuf parameters
it could lead to the construction of invalid mutators.

At the point where an enum type is reached we now use a simplified
factory instead that only deals with repeated or nullable enums.