feat: add basic mutator runtime
Propagate a new class MutatorRuntime to every mutator. For now the
class only contains the fuzz test method, necessary for implementation
of @DictionaryProvider annotation, in order to extract thunks
returning Stream<?>.
This class can be extended in the future to contain the table of
recent compares, etc. to improve fuzzing performance.