fix: fuzz test invocation with referring protobufs
When constructing a mutator for a protobuf message the mutator is cached based
on the protobuf descriptor. This led to issues for fuzz tests that take two
protobuf messages where one message is referring to the other.
While building the mutator for the top level message nested messages are
constructed with the type `DynamicMessage` which lead to 'argument type
mismatch' exceptions when invoking the fuzz test method.
This is fixed by including the message class in the mutator cache key.