fix: assertion failure when fuzzing proto strings
When parsing protobuf messages from libfuzzer bytes in the protobuf
mutator we don't enforce length limits. This could lead to
maxSizeIncrease < 0 situations if the e.g. a String read by the protobuf
mutator exceeds the size limit.
To guard against similar cases we now enforce the size constraints for
the byte array that is being mutated before performing the mutation.