fix: instrumention of nested records, and records with annotated fields
Using ClassWriter without ClassReader in asm, seems to produce
corrupted bytecode for nested records and records with annotated fields.

Prior to this fix, a segfault was triggered for nested records that
use Jazzer annotations when Jazzer was trying to access data on record
components, when trying to create a record mutator.

In addition, for non-nested records with annotated fields as for example:
record Address(byte @WithLength(max=10) [] data) {}
no suitable mutator could be found.